refactor to reuse common deploy and validate logic
