+++ packages/vscode-extension/src/service/TranslationKeys.ts
@@ -230,3 +230,7 @@ export const UPDATE_OBJECT = {
 	PROCESS_CANCELED: 'UPDATE_OBJECT_PROCESS_CANCELED',
 	UPDATING: 'UPDATE_OBJECT_UPDATING',
 };
+
+export const VALIDATE = {
+	VALIDATING: 'VALIDATE_VALIDATING'
+}
